+++ src/Natives.cpp
@@ -38,7 +38,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_RayCastLine(AMX *amx, cell *params)
 	btVector3 Start = btVector3(btScalar(amx_ctof(params[1]) + 0.00001), btScalar(amx_ctof(params[2]) + 0.00001), btScalar(amx_ctof(params[3]) + 0.00001));
 	btVector3 End = btVector3(btScalar(amx_ctof(params[4])), btScalar(amx_ctof(params[5])), btScalar(amx_ctof(params[6])));
 	btVector3 Result;
-	uint16_t Model = 0;
+	int32_t Model = 0;
 
 	if (collisionWorld->performRayTest(Start, End, Result, Model))
 	{
@@ -121,7 +121,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_RayCastLineEx(AMX *amx, cell *params)
 	btVector3 Position;
 	btQuaternion Rotation;
 	btVector3 Result;
-	uint16_t Model = 0;
+	int32_t Model = 0;
 
 	if (collisionWorld->performRayTestEx(Start, End, Result, Rotation, Position, Model))
 	{
@@ -164,7 +164,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_RayCastLineAngle(AMX *amx, cell *para
 	btScalar RX;
 	btScalar RY;
 	btScalar RZ;
-	uint16_t model = 0;
+	int32_t model = 0;
 
 	if (collisionWorld->performRayTestAngle(Start, End, Result, RX, RY, RZ, model))
 	{
@@ -202,7 +202,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_RayCastLineAngleEx(AMX *amx, cell *pa
 	btScalar RX;
 	btScalar RY;
 	btScalar RZ;
-	uint16_t Model = 0;
+	int32_t Model = 0;
 
 	if (collisionWorld->performRayTestAngleEx(Start, End, Result, RX, RY, RZ, ObjectRotation, ObjectPosition, Model))
 	{
@@ -304,7 +304,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_CreateObject(AMX *amx, cell *params)
 		return -1;
 	}
 	
-	uint16_t modelid = static_cast<uint16_t>(params[1]);
+	int32_t modelid = static_cast<int32_t>(params[1]);
 	uint16_t addtomanager = static_cast<uint16_t>(params[8]);
 
 	if (collisionWorld->getModelRef(modelid) != 65535)
@@ -424,7 +424,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_SetObjectRot(AMX *amx, cell *params)
 
 cell AMX_NATIVE_CALL ColAndreasNatives::CA_GetModelBoundingSphere(AMX *amx, cell *params)
 {
-	uint16_t modelid = static_cast<uint16_t>(params[1]);
+	int32_t modelid = static_cast<int32_t>(params[1]);
 	
 	if (modelid >= 0 && modelid < 20000)
 	{
@@ -453,7 +453,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_GetModelBoundingSphere(AMX *amx, cell
 
 cell AMX_NATIVE_CALL ColAndreasNatives::CA_GetModelBoundingBox(AMX *amx, cell *params)
 {
-	uint16_t modelid = static_cast<uint16_t>(params[1]);
+	int32_t modelid = static_cast<int32_t>(params[1]);
 	
 	if (modelid >= 0 && modelid < 20000)
 	{
@@ -507,7 +507,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_RayCastReflectionVector(AMX *amx, cel
 	btVector3 End = btVector3(btScalar(amx_ctof(params[4])), btScalar(amx_ctof(params[5])), btScalar(amx_ctof(params[6])));
 	btVector3 Result;
 	btVector3 Position;
-	uint16_t model;
+	int32_t model;
 
 	if (collisionWorld->performRayTestReflection(Start, End, Position, Result, model))
 	{
@@ -545,7 +545,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_RayCastLineNormal(AMX *amx, cell *par
 	btVector3 End = btVector3(btScalar(amx_ctof(params[4])), btScalar(amx_ctof(params[5])), btScalar(amx_ctof(params[6])));
 	btVector3 Result;
 	btVector3 Position;
-	uint16_t model;
+	int32_t model;
 
 	if (collisionWorld->performRayTestNormal(Start, End, Position, Result, model))
 	{
@@ -573,7 +573,7 @@ cell AMX_NATIVE_CALL ColAndreasNatives::CA_RayCastLineNormal(AMX *amx, cell *par
 
 cell AMX_NATIVE_CALL ColAndreasNatives::CA_ContactTest(AMX *amx, cell *params)
 {
-	uint16_t modelid = static_cast<uint16_t>(params[1]);
+	int32_t modelid = static_cast<int32_t>(params[1]);
 	
 	btVector3 position = btVector3(amx_ctof(params[2]), amx_ctof(params[3]), amx_ctof(params[4]));
 	btVector3 rotation = btVector3(amx_ctof(params[5]), amx_ctof(params[6]), amx_ctof(params[7]));
